requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
py==1.11.0
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
tomli==2.0.1
urllib3==2.0.4

mysql-connector-python~=8.1.0
python-dotenv~=1.0.0